# Pre-parsed path put/remove and performance improvements