--- a/src/pljson.type.decl.sql
+++ b/src/pljson.type.decl.sql
@@ -93,7 +93,9 @@ create or replace type pljson force under pljson_element (
 
   /* map functions */
   member function get_keys return pljson_list,
-  member function get_values return pljson_list
+  member function get_values return pljson_list,
+
+  overriding member function internal_path_put(self in out nocopy pljson, path pljson_path, elem pljson_element, path_position pls_integer) return boolean
 ) not final;
 /
 show err