--- a/src/pljson_ext.impl.sql
+++ b/src/pljson_ext.impl.sql
@@ -287,14 +287,12 @@ create or replace package body pljson_ext as
 
     return ret;
   end parsePath;
-  
-  --JSON Path getters
-  function get_json_element(obj pljson, v_path varchar2, base number default 1) return pljson_element as
-    path pljson_list;
+
+  --JSON pre-parsed path getters
+  function get_json_element(obj pljson, path pljson_list) return pljson_element as
     ret pljson_element;
     o pljson; l pljson_list;
   begin
-    path := parsePath(v_path, base);
     ret := obj;
     if (path.count = 0) then return ret; end if;
 
@@ -336,7 +334,91 @@ create or replace package body pljson_ext as
     when others then return null;
   end get_json_element;
 
+  function get_string(obj pljson, path pljson_list) return varchar2 as
+    temp pljson_element;
+  begin
+    temp := get_json_element(obj, path);
+    if (temp is null or not temp.is_string()) then
+      return null;
+    else
+      return temp.get_string();
+    end if;
+  end;
+
+  function get_number(obj pljson, path pljson_list) return number as
+    temp pljson_element;
+  begin
+    temp := get_json_element(obj, path);
+    if (temp is null or not temp.is_number()) then
+      return null;
+    else
+      return temp.get_number();
+    end if;
+  end;
+
+  function get_double(obj pljson, path pljson_list) return binary_double as
+    temp pljson_element;
+  begin
+    temp := get_json_element(obj, path);
+    if (temp is null or not temp.is_number()) then
+      return null;
+    else
+      return temp.get_double();
+    end if;
+  end;
+
+  function get_json(obj pljson, path pljson_list) return pljson as
+    temp pljson_element;
+  begin
+    temp := get_json_element(obj, path);
+    if (temp is null or not temp.is_object()) then
+      return null;
+    else
+      return treat(temp as pljson);
+    end if;
+  end;
+
+  function get_json_list(obj pljson, path pljson_list) return pljson_list as
+    temp pljson_element;
+  begin
+    temp := get_json_element(obj, path);
+    if (temp is null or not temp.is_array()) then
+      return null;
+    else
+      return treat(temp as pljson_list);
+    end if;
+  end;
+
+  function get_bool(obj pljson, path pljson_list) return boolean as
+    temp pljson_element;
+  begin
+    temp := get_json_element(obj, path);
+    if (temp is null or not temp.is_bool()) then
+      return null;
+    else
+      return temp.get_bool();
+    end if;
+  end;
+
+  function get_date(obj pljson, path pljson_list) return date as
+    temp pljson_element;
+  begin
+    temp := get_json_element(obj, path);
+    if (temp is null or not is_date(temp)) then
+      return null;
+    else
+      return pljson_ext.to_date(temp);
+    end if;
+  end;
+
   --JSON Path getters
+  function get_json_element(obj pljson, v_path varchar2, base number default 1) return pljson_element as
+    path pljson_list;
+  begin
+    path := parsePath(v_path, base);
+    return get_json_element(obj, path);
+  end get_json_element;
+
   function get_string(obj pljson, path varchar2, base number default 1) return varchar2 as
     temp pljson_element;
   begin