--- a/src/pljson_list.type.decl.sql
+++ b/src/pljson_list.type.decl.sql
@@ -92,7 +92,9 @@ create or replace type pljson_list force under pljson_element (
   member procedure path_put(self in out nocopy pljson_list, json_path varchar2, elem pljson_list, base number default 1),
 
   /* json path_remove */
-  member procedure path_remove(self in out nocopy pljson_list, json_path varchar2, base number default 1)
+  member procedure path_remove(self in out nocopy pljson_list, json_path varchar2, base number default 1),
+
+  overriding member function internal_path_put(self in out nocopy pljson_list, path pljson_path, elem pljson_element, path_position pls_integer) return boolean
 ) not final;
 /
 show err