@asfernandes 2020-07-30
@dsnz the pull request has sequential commits which make easy to incrementally apply/revert. But let's see if we are understanding well each other.

The first set of commits improves performance when using the new API (pre-parsed paths).

The second set of commits (change to `put_internal` - actually `put_internal_preparsed`) improves both the string and pre-parsed path APIs.

So you want that I make two code paths where the string-based path goes to the old code and the pre-parsed one goes to the new code?