@asfernandes 2020-07-30
> and what do you think of renaming `internal_path_put` to some other name to match more closely the existing `put_internal` (something like `path_put_internal` or `put_internal_path` perhaps) ?

I prefixed it with "internal" to make clear to users it should not be used by them. But if you prefer I can rename it.