@asfernandes 2020-07-30
> Hi
> 
> `procedure put_internal(obj in out nocopy pljson, v_path varchar2, elem pljson_element, base number) as...`
> 
> is not used in new procedures
> your new procedures use only the new `put_internal_path` (after the rename)
> 

Yes, sure. This is why I asked you to confirm about two code paths.

> the existing string based procedures are not changed but affected indirectly only by changing the existing `put_internal`
> I can't say if changing the existing one to use the new ```put_internal_path`` that it will be speedier,
> probably not at all (because the same operations are done in the end and only for one time so the pre-parsing does not speed anything) but it makes for **code commonality** which is useful
> 

It's true that parsing the string path and them calling the function with deals with pre-parsed path does not improve performance per se, but that is not the whole picture. Using `put_internal_path` is what makes the close to 50% performance improvement in the string-based path API.

> however as I said the code differences are many and I can't check the correct working so fast and just by inspection, there are many **subtle points**
> that's why I ask that you retract **only this** and put it in a separate pull request
> 
> it's a small change in your pull request now but ensures that we check this thing better in the next days
> 

Sure, I understand that. The routine was difficult to be replicated in the `put_internal_path` with the same behaviors and could have problems. I'll make the string-based path using the same old path (and hence having no performance improvement) in this pull request and create additional pull request reverting that (i.e. maintaining the behavior as the current version of this PR).

> yes please change the name to `put_internal_path` so their names match as their functions match
> and you can put a comment with big letters before stating that this is internal procedure and not part of the API (put this comment in both declarations and implementations)

Ok.