@asfernandes 2020-07-30
> great 👍
> I'm waiting for the updated pull request
> and the new pull request that will follow

Done.